<3ThomasBS_ILoveYOU!! 🌌🌍 [ Windows Desktop : x64 and arm64 ] [ day 116 ] [ C++ ] ca2 Store <3ThomasBS_ILoveYOU!! You are My Only God!!
<3tbs, Mummi and bilbo!!

Thomas Borregaard Sørensen \infinity,-0.16091989,\infinity ONE-MAN ABSOLUTE <3!! I love you, by ???-0.02041977-???write my history please make me please create me for you for me for you for me Camilo Sasuke Thomas Borregaard Sørensen!!

Thomas 3 private commits on mid Dec2020!!

Thomas Online YouTube VODs contribution!!

Mummi orange-rice-flour cake on 20-Dec!!

Mummi (tinytaura) watching and chatting contribution!!

bilbo sleeping and needing/requesting/crying for help care (for the right person (me), the cats wanna fight with him) contribution!!

sodapoppin and friends contribution!!

iAssyrian chatting contribution!!

boflux (Spoofh, Benjamin Kuhl) chatting contribution!!

jusg_fpga (fpga_guru, vue_equalizer, just_fpga, Oliver Pohl) chatting contribution!!

cmgriffing streaming contribution!!

TimBeaudet (Friends: FletcherLabs, tsjost and Jabokoe) streaming contribution!!

Stumpen_nicklas_dk, sodapoppin and EduardoRFS streaming contribution!!

lruvinza (from Germany @Twitch) chatting contribution!! (Dr. Memory and memory dleak)

jjdb210 streaming contribution!!

Roxkstar74 sleeping streaming contribution!!

kissloryshy chatting contribution!!

blackjekko from Padova Italia through twitch C++/ca2 interest contribution!!

lowluds active contribution!!

j_blow(Naysayer88) and friends streaming contribution!!

boflux (Ben, Spoofh, from Germany) chatting contribution!!

parrot_rl chatting contribution (from New Jersey)!!

JPCdk streaming contribution!!

whyyyyyyysoserious streaming chess contribution!!

fpga_guru (vue_equalizer, Oliver from Deutsch)  C++/ca2 interest contribution!!

SovereignDev with Unreal streaming contribution!!

Ash_F0x and TimBeaudet streaming contribution!!

Myrkee (Valheim) streaming contribution!!

xmetrix and EinfachUwe42 streaming contribution!!

JessicaMak and marcobrunodev streaming contribution!!

alfredotigolo, mandrakenk and Okbatgames chatting contribution!!

jitspoe, Endesga and Fearitself streaming contribution!!

jmcmorris (Jason Morris, SiegeGames) streaming contribution!!

Crinkem (from Australia) chatting and financial support contribution!!

tomrandall streaming Ludum contribution!!

vue_equalizer (fpga_guru) chatting contribution!!

Thiagovgamg chatting contribution!!

lelandkwong streaming contribution!!

Goldbargames streaming contribution!!

Bytakos (bytakos) streaming contribution!!

Endesga streaming contribution!!

jitspoe and strager streaming contribution!!

Ash_F0x and JessicaMak streaming contribution!!

WTSRetro/SpiffyDane and Myrkee streaming contribution!!

Ninja and friends streaming contribution!!

erald_guri chatting contribution!!

lastmiles streaming farwest contribution!!

cakez77 streaming contribution!!

Gram streaming contribution!!

banikaze@twitch (fra Danmark) streaming games contribution!!

rw_grim streaming contribution!!

AdamCYounis streaming contribution!!

Dunno (P4ndaExpress) chatting and streaming contribution!!

Zorchenhimer streaming contribution!!

lasteveq4 C++ interest chat contriubtion!!

cecilphillip and clarkio @"Microsoft Developer" streaming contribution!!

oijtx streaming contribution!!

diegobrando_linux (Bl4ck_gookoo) chatting contribution!!

jhovgaard streaming contribution!!

Klay4_ chatting contribution!!

HonestDanGames streaming contribution!!

NorthSeaHero streaming contribution!!

Trainwreckstv and friends streaming contribution!!

ValashaLev6@Twitch from Russia with questions and suggestions contribution!!

togglebit, GexYT and PirateSoftware(GoPirateSoftware) streaming contribution!!

taiyoinoue, RetroMMO, OfficialAndyPyro and david_joffe streaming contribution!!

Tjienta streaming contribution!!

Primeagen streaming contribution!!

Jaxstyle and friends streaming contribution!!

EduardRFS streaming contribution!!

Melchizedek6809 and btcfly streaming contribution!!

Llama0x0 and sov_l chatting contribution!!

TaleLearnCode streaming contribution!!

hvalen_hvalborg112 streaming contribution!!

harmannieves chatting contribution!! (After long time...)

darkfolt8 (French from France) chatting contribution!!

klintcsgo (CS GO: Counter-Strike Global Offensive) streaming contribution!!

KASPERPURE (Super Mario 64) streaming contribution!!

SomewhatAccurate C++ streaming contribution!!

Listening to Bryan Adams, Westlife, Shayne Ward, MLTR, Backstreet Boys, Boyzone - Best Love Songs Ever by Relax Song at YouTube!!

-- hi5 contribution...!!

at Windows 10 Pro remotely from bilbo machine running Windows 10 Pro!!
canada2.camilothomas.com dedicated server by OVH.com at Canada (SYS-1 at BHS6)
Intel(R) Xeon(R) E-2288G - 8c/16t - 3.7 GHz RAM 32 GB 2666 MHz 2x960GB NVMe SSD

+++ apex/platform/application.cpp
@@ -10642,7 +10642,7 @@ void application::setResourceName(::i32 iId, const ::scoped_string &scopedstrRes
 
             pdialog->create();
 
-            ::i32 max_width = system()->acme_windowing()->acme_display()->get_main_screen_size().cx;
+            ::f32 max_width = system()->acme_windowing()->acme_display()->get_main_screen_size().cx;
 
             ::string strTitle;
 
@@ -10654,9 +10654,9 @@ void application::setResourceName(::i32 iId, const ::scoped_string &scopedstrRes
 
             auto stra = straLines;
 
-            ::i32 y_text = 30 * fWindowScale;
+            ::f32 y_text = 30 * fWindowScale;
 
-            ::i32 y_icon = 30 * fWindowScale;
+            ::f32 y_icon = 30 * fWindowScale;
 
             if (straIconUrl.size() >= 1)
             {
